CLOUD-61408 Remove unnecessary (un)boxings

+++ cloud-gcp/src/main/java/com/sequenceiq/cloudbreak/cloud/gcp/util/GcpStackUtil.java
@@ -96,7 +96,7 @@ public final class GcpStackUtil {
             throw new Exception(errorMessage);
         } else {
             Integer progress = operation.getProgress();
-            return progress.intValue() == FINISHED;
+            return progress == FINISHED;
         }
     }
